'a'
In the line above a is our variable. The number 14 is the value we'd like to store in 'a'. Finally, the word const is a keyword used to declare 'a' as a constant value. We can access the value 14 in a future line of code by referring to the variable a